[RENAME] protocol파일 따로 분리
